resolvendo erro IEEE 754: uso de BigDecimal

+++ src/test/java/com/simplebank/service/SimpleBankTests.java
@@ -9,12 +9,12 @@ import static org.assertj.core.api.Assertions.*;
 public class SimpleBankTests {
     @Test
     void dadaContaCorrenteAComSaldo_eOutraContaCorrenteBSaldo_quandoAtransferirValorB_entaoContasSaldosComputados() {
-        var contaCorrenteA = new ContaCorrente(0.3);
-        var contaCorrenteB = new ContaCorrente(0);
+        var contaCorrenteA = new ContaCorrente(BigDecimal.valueOf(0.3));
+        var contaCorrenteB = new ContaCorrente(BigDecimal.valueOf(0));
 
-        contaCorrenteA.transfere(contaCorrenteB, 0.1);
+        contaCorrenteA.transfere(contaCorrenteB, BigDecimal.valueOf(0.1));
 
-        assertThat(contaCorrenteA.getSaldo()).isEqualTo(0.2);
-        assertThat(contaCorrenteB.getSaldo()).isEqualTo(0.1);
+        assertThat(contaCorrenteA.getSaldo()).isEqualTo(BigDecimal.valueOf(0.2));
+        assertThat(contaCorrenteB.getSaldo()).isEqualTo(BigDecimal.valueOf(0.1));
     }
 }
